refactor: ViewResolver 분리
