+++ app/src/main/java/nextstep/jwp/http/Controller.java
@@ -1,15 +1,11 @@
-package nextstep.jwp;
+package nextstep.jwp.http;
 
 import static nextstep.jwp.http.HttpResponse.found;
-import static nextstep.jwp.http.HttpResponse.ok;
 import static nextstep.jwp.http.HttpResponse.redirect;
 import static nextstep.jwp.http.HttpResponse.unauthorized;
+import static nextstep.jwp.http.ViewResolver.resolveView;
 
-import java.io.File;
 import java.io.IOException;
-import java.net.URL;
-import java.nio.file.Files;
-import java.nio.file.Path;
 import java.util.HashMap;
 import java.util.Map;
 import nextstep.jwp.db.InMemoryUserRepository;
@@ -18,7 +14,7 @@ import nextstep.jwp.model.User;
 public class Controller {
 
     public String login() throws IOException {
-        return view("/login.html");
+        return resolveView("login");
     }
 
     public String login(String requestBody) {
@@ -33,7 +29,7 @@ public class Controller {
     }
 
     public String register() throws IOException {
-        return view("/register.html");
+        return resolveView("register");
     }
 
     public String register(String requestBody) throws IOException {
@@ -55,13 +51,5 @@ public class Controller {
             queryMap.put(token[0], token[1]);
         }
         return queryMap;
-    }
-
-    private String view(String uri) throws IOException {
-        final URL resource = getClass().getClassLoader().getResource("static" + uri);
-        final Path path = new File(resource.getPath()).toPath();
-        String responseBody = Files.readString(path);
-
-        return ok("text/html", responseBody);
     }
 }
